+++ .idea/dictionaries/abdur_raqeeb.xml
@@ -1,6 +1,7 @@
 <component name="ProjectDictionaryState">
   <dictionary name="abdur-raqeeb">
     <words>
+      <w>axios</w>
       <w>linkedin</w>
     </words>
   </dictionary>
