+++ src/components/atoms/FooterContactItem.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const FooterContactItem = ({ icon, text }) => (
+    <section>
+        {icon}
+        <p>{text}</p>
+    </section>
+);
+
+export default FooterContactItem;
