+++ src/components/atoms/Spinner.jsx
@@ -1,4 +1,4 @@
-import spinner from '../assets/spinner.gif';
+import spinner from '../../assets/spinner.gif';
 
 const Spinner = () => {
   return (
